// packages/schemas/features/blocks/inputs/constants.ts
export const defaultButtonLabel = 'Send'

export enum InputBlockType {
  TEXT = 'text input',
  NUMBER = 'number input',
  EMAIL = 'email input',
  URL = 'url input',
  DATE = 'date input',
  PHONE = 'phone number input',
  CHOICE = 'buttons input',
  PAYMENT = 'stripe',
  RATING = 'rating input',
  FILE = 'file input',
}